Final style refactoring
Todos for after merge (also left these as code comments):
- Abstraction in `useSchedule`
- Generic `Picker` component; see `PaletteClassFilter`, `PriorityPicker`
- Think about dot abstraction; see `PaletteClassFilter`
