+++ client/src/components/classes/UpcomingFullCalendar.tsx
@@ -1,14 +1,13 @@
 import { useState } from 'react';
 import moment from 'moment';
 import { GenericCalendar } from '../schedule/DateSelector';
-import { ActiveDayState } from './upcoming/Assignments';
+import { ActiveDayState } from './UpcomingAssignments';
 import { DateRangeProps } from '../schedule/DateSelector';
 
 
 // Side calendar for large screens on Upcoming
 // Doubles as a date range filter for the assignments, and for a quick visualization of due dates
 export default function UpcomingFullCalendar(props: DateRangeProps & ActiveDayState) {
-
     const { start, end, setStart, setEnd, activeDay } = props;
     const endInclusive = end.clone().subtract(1, 'days');
 
@@ -27,7 +26,7 @@ export default function UpcomingFullCalendar(props: DateRangeProps & ActiveDaySt
                     if (selecting === 'S') {
                         if (day.isBefore(end)) setStart(day);
                     } else {
-                        const ex = moment(day); ex.add(1, 'days');
+                        const ex = moment(day).add(1, 'days');
                         if (ex.isAfter(start)) {
                             setEnd(ex);
                         }
